Move special iterators to a separate file
